Fixed Permissions bug in Deck deletion, Updated DeckCard to associate with User

+++ app/controllers/decks_controller.rb
@@ -97,6 +97,8 @@ class DecksController < ApplicationController
 
         def destroy
             deck = find_deck
+            return render json: {error: "Not Authorized"}, status: :unauthorized unless deck.user_id == session[:user_id]
+
             deck.destroy
             :no_head
         end
